+++ bin/test.sh
@@ -1,3 +1,3 @@
 #!/usr/bin/env bash
 
-vendor/bin/phpspec run -f dot && vendor/bin/phpunit && vendor/bin/php-cs-fixer fix --dry-run --config=sf23 .
+vendor/bin/phpspec --no-interaction run --format=dot && vendor/bin/phpunit && vendor/bin/php-cs-fixer fix --dry-run --config=sf23 .